Updated samples to use kendra retriever from Langchain
