--- aws_langchain/__init__.py
@@ -1,1 +1,0 @@
-"""Classes to work with AWS Kendra and Bedrock LLMs"""
