+++ kendra_retriever_samples/images/user-icon.png
