+++ kendra_retriever_samples/kendra-docs-index.yaml
