+++ kendra_retriever_samples/kendra_chat_open_ai.py
@@ -1,4 +1,4 @@
-from aws_langchain.kendra_index_retriever import KendraIndexRetriever
+from langchain.retrievers import AmazonKendraRetriever
 from langchain.chains import ConversationalRetrievalChain
 from langchain.prompts import PromptTemplate
 from langchain import OpenAI
@@ -13,9 +13,7 @@ def build_chain():
 
   llm = OpenAI(batch_size=5, temperature=0, max_tokens=300)
       
-  retriever = KendraIndexRetriever(kendraindex=kendra_index_id, 
-      awsregion=region, 
-      return_source_documents=True)
+  retriever = AmazonKendraRetriever(index_id=kendra_index_id)
 
   prompt_template = """
   The following is a friendly conversation between a human and an AI. 
@@ -23,14 +21,30 @@ def build_chain():
   If the AI does not know the answer to a question, it truthfully says it 
   does not know.
   {context}
-  Instruction: Based on the above documents, provide a detailed answer for, {question} Answer "don't know" if not present in the document. Solution:
-  """
+  Instruction: Based on the above documents, provide a detailed answer for, {question} Answer "don't know" 
+  if not present in the document. 
+  Solution:"""
   PROMPT = PromptTemplate(
       template=prompt_template, input_variables=["context", "question"]
   )
 
-  return ConversationalRetrievalChain.from_llm(llm=llm, retriever=retriever, qa_prompt=PROMPT, return_source_documents=True)
-  
+  condense_qa_template = """
+  Given the following conversation and a follow up question, rephrase the follow up question 
+  to be a standalone question.
+
+  Chat History:
+  {chat_history}
+  Follow Up Input: {question}
+  Standalone question:"""
+  standalone_question_prompt = PromptTemplate.from_template(condense_qa_template)
+
+  qa = ConversationalRetrievalChain.from_llm(
+        llm=llm, 
+        retriever=retriever, 
+        condense_question_prompt=standalone_question_prompt, 
+        return_source_documents=True, 
+        combine_docs_chain_kwargs={"prompt":PROMPT})
+  return qa
 
 def run_chain(chain, prompt: str, history=[]):
   return chain({"question": prompt, "chat_history": history})
