+++ kendra_retriever_samples/kendra_retriever_anthropic.py
@@ -1,0 +1,62 @@
+from langchain.retrievers import AmazonKendraRetriever
+from langchain.chains import RetrievalQA
+from langchain.prompts import PromptTemplate
+from langchain.chat_models import ChatAnthropic as Anthropic
+import os
+
+
+def build_chain():
+  ANTHROPIC_API_KEY = os.environ["ANTHROPIC_API_KEY"]
+  region = os.environ["AWS_REGION"]
+  kendra_index_id = os.environ["KENDRA_INDEX_ID"]
+
+  llm = Anthropic(temperature=0, anthropic_api_key=ANTHROPIC_API_KEY)
+        
+  retriever = AmazonKendraRetriever(index_id=kendra_index_id)
+
+  prompt_template = """
+
+  Human: This is a friendly conversation between a human and an AI. 
+  The AI is talkative and provides specific details from its context but limits it to 240 tokens.
+  If the AI does not know the answer to a question, it truthfully says it 
+  does not know.
+
+  Assistant: OK, got it, I'll be a talkative truthful AI assistant.
+
+  Human: Here are a few documents in <documents> tags:
+  <documents>
+  {context}
+  </documents>
+  Based on the above documents, provide a detailed answer for, {question} Answer "don't know" 
+  if not present in the document. 
+
+  Assistant:"""
+
+  PROMPT = PromptTemplate(
+      template=prompt_template, input_variables=["context", "question"]
+  )
+  chain_type_kwargs = {"prompt": PROMPT}
+  return RetrievalQA.from_chain_type(
+      llm, 
+      chain_type="stuff", 
+      retriever=retriever, 
+      chain_type_kwargs=chain_type_kwargs,
+      return_source_documents=True
+  )
+
+def run_chain(chain, prompt: str, history=[]):
+    result = chain(prompt)
+    # To make it compatible with chat samples
+    return {
+        "answer": result['result'],
+        "source_documents": result['source_documents']
+    }
+
+if __name__ == "__main__":
+    chain = build_chain()
+    result = run_chain(chain, "What's SageMaker?")
+    print(result['answer'])
+    if 'source_documents' in result:
+        print('Sources:')
+        for d in result['source_documents']:
+          print(d.metadata['source'])
