+++ kendra_retriever_samples/kendra_retriever_flan_xxl.py
@@ -1,9 +1,9 @@
-from aws_langchain.kendra_index_retriever import KendraIndexRetriever
+from langchain.retrievers import AmazonKendraRetriever
 from langchain.chains import RetrievalQA
 from langchain import OpenAI
 from langchain.prompts import PromptTemplate
 from langchain import SagemakerEndpoint
-from langchain.llms.sagemaker_endpoint import ContentHandlerBase
+from langchain.llms.sagemaker_endpoint import LLMContentHandler
 import json
 import os
 
@@ -13,7 +13,7 @@ def build_chain():
     kendra_index_id = os.environ["KENDRA_INDEX_ID"]
     endpoint_name = os.environ["FLAN_XXL_ENDPOINT"]
 
-    class ContentHandler(ContentHandlerBase):
+    class ContentHandler(LLMContentHandler):
         content_type = "application/json"
         accepts = "application/json"
 
@@ -34,9 +34,7 @@ def build_chain():
             content_handler=content_handler
         )
 
-    retriever = KendraIndexRetriever(kendraindex=kendra_index_id, 
-        awsregion=region, 
-        return_source_documents=True)
+    retriever = AmazonKendraRetriever(index_id=kendra_index_id)
 
     prompt_template = """
     The following is a friendly conversation between a human and an AI. 
@@ -44,8 +42,9 @@ def build_chain():
     If the AI does not know the answer to a question, it truthfully says it 
     does not know.
     {context}
-    Instruction: Based on the above documents, provide a detailed answer for, {question} Answer "don't know" if not present in the document. Solution:
-    """
+    Instruction: Based on the above documents, provide a detailed answer for, {question} Answer "don't know" 
+    if not present in the document. 
+    Solution:"""
     PROMPT = PromptTemplate(
         template=prompt_template, input_variables=["context", "question"]
     )
