+++ kendra_retriever_samples/requirements.txt
@@ -1,0 +1,5 @@
+langchain>=0.0.213
+boto3>=1.26.159
+openai
+anthropic
+streamlit
